% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtp_selector.R
\name{dtp_projected_dose}
\alias{dtp_projected_dose}
\title{Projected dose after waiting.}
\usage{
dtp_projected_dose(x, ...)
}
\arguments{
\item{x}{A \code{dtp_selector} object.}

\item{...}{Additional arguments (unused).}
}
\value{
Integer dose level, or \code{NA} if waiting has no benefit.
}
\description{
Projected dose after waiting.
}
