% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtp_selector.R
\name{dtp_should_wait}
\alias{dtp_should_wait}
\title{Should the trialist wait before dosing the next cohort?}
\usage{
dtp_should_wait(x, ...)
}
\arguments{
\item{x}{A \code{dtp_selector} object.}

\item{...}{Additional arguments (unused).}
}
\value{
Logical scalar.
}
\description{
Returns \code{TRUE} if the DTP projection found that waiting could lead to
a higher dose recommendation within \code{t_max}.
}
