% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtptite-package.R
\docType{package}
\name{dtptite-package}
\alias{dtptite}
\alias{dtptite-package}
\title{dtptite: Dose Transition Pathways for Time-to-Event Dose-Finding Designs}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Implements the Dose Transition Pathways (DTP) look-ahead strategy for time-to-event phase I dose-finding clinical trials with delayed or pending toxicity outcomes. Provides a DTP decorator for the 'escalation' framework, supporting both TITE-CRM (model-based) and TITE-BOIN (model-assisted) base designs.
}
\section{Funding}{

Xiaoran Lai, NIHR Development and Skills Enhancement Award (DSE),
NIHR500663, is funded by the NIHR for this research project. The views
expressed in this publication are those of the author(s) and not
necessarily those of the NIHR, NHS or the UK Department of Health and
Social Care.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ICR-EPAT/dtptite}
  \item Report bugs at \url{https://github.com/ICR-EPAT/dtptite/issues}
}

}
\author{
\strong{Maintainer}: Xiaoran Lai \email{xiaoran.lai@icr.ac.uk}

Other contributors:
\itemize{
  \item The Institute of Cancer Research [copyright holder]
  \item National Institute for Health and Care Research [funder]
}

}
\keyword{internal}
